v0.3.1 Source

+++ TownOfUsReworked/BetterMaps/Airship/Repositioning.cs
@@ -7,11 +7,11 @@
         {
             var AdminTable = UObject.FindObjectOfType<MapConsole>();
 
-            if ((byte)CustomGameOptions.MoveAdmin != 0)
+            if (CustomGameOptions.MoveAdmin != 0)
             {
                 var MapFloating = GameObject.Find("Cockpit/cockpit_mapfloating");
 
-                if ((byte)CustomGameOptions.MoveAdmin == 1)
+                if ((int)CustomGameOptions.MoveAdmin == 1)
                 {
                     AdminTable.transform.position = new(-17.269f, 1.375f);
                     AdminTable.transform.rotation = Quaternion.Euler(new(0, 0, 350.316f));
@@ -21,9 +21,9 @@
                     MapFloating.transform.rotation = Quaternion.Euler(new(0, 0, 350));
                     MapFloating.transform.localScale = new(1, 1, 1);
                 }
-                else if ((byte)CustomGameOptions.MoveAdmin == 2)
+                else if ((int)CustomGameOptions.MoveAdmin == 2)
                 {
-                    // New Admin
+                    //New Admin
                     AdminTable.transform.position = new(5.078f, 3.4f, 1);
                     AdminTable.transform.rotation = Quaternion.Euler(new(0, 0, 76.1f));
                     AdminTable.transform.localScale = new(1.200f, 1.700f, 1);
@@ -31,11 +31,11 @@
                 }
             }
 
-            if ((byte)CustomGameOptions.MoveElectrical != 0)
+            if (CustomGameOptions.MoveElectrical != 0)
             {
                 var Electrical = GameObject.Find("GapRoom/task_lightssabotage (gap)");
 
-                if ((byte)CustomGameOptions.MoveElectrical == 1)
+                if ((int)CustomGameOptions.MoveElectrical == 1)
                 {
                     Electrical.transform.position = new(-8.818f, 13.184f);
                     Electrical.transform.localScale = new(0.909f, 0.818f, 1);
@@ -46,7 +46,7 @@
                     SupportElectrical.transform.position = new(-8.792f, 13.242f);
                     SupportElectrical.transform.localScale = new(1, 1, 1);
                 }
-                else if ((byte)CustomGameOptions.MoveElectrical == 2)
+                else if ((int)CustomGameOptions.MoveElectrical == 2)
                     Electrical.transform.position = new(19.339f, -3.665f);
             }
